rag/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-groq
langchain-huggingface
faiss-cpu
youtube-transcript-api
yt-dlp
Pillow
faster-whisper
python-dotenv
pydantic
pydantic-settings
pytest
ruff
httpx
